исправил дз

+++ src/main/java/pro/sky/java/course2/calculator/service/CalculatedService.java
@@ -2,7 +2,10 @@ package pro.sky.java.course2.calculator.service;
 
 public interface CalculatedService {
     int plus(int x, int y);
+
     int minus(int x, int y);
+
     int multiply(int x, int y);
-    int divide(int x, int y);
+
+    double divide(int x, int y);
 }
